+++ src/api/configure.ts
@@ -1,10 +1,14 @@
 import { globalState, isolateGlobalState } from "../core/globalstate"
+import { reserveArrayBuffer } from "../types/observablearray"
+import { setReactionScheduler } from "../core/reaction"
 
 export function configure(options: {
     enforceActions?: boolean
     computedRequiresReaction?: boolean
     isolateGlobalState?: boolean
     disableErrorBoundaries?: boolean
+    arrayBuffer?: number
+    reactionScheduler: (f: () => void) => void
 }): void {
     if (options.enforceActions !== undefined) {
         globalState.enforceActions = !!options.enforceActions
@@ -22,5 +26,11 @@ export function configure(options: {
                 "WARNING: Debug feature only. MobX will NOT recover from errors if this is on."
             )
         globalState.disableErrorBoundaries = !!options.disableErrorBoundaries
+    }
+    if (typeof options.arrayBuffer === "number") {
+        reserveArrayBuffer(options.arrayBuffer)
+    }
+    if (options.reactionScheduler) {
+        setReactionScheduler(options.reactionScheduler)
     }
 }