+++ src/mobx.ts
@@ -108,9 +108,7 @@ export { getDebugName, getAtom, getAdministration as _getAdministration } from "
 export { allowStateChanges as _allowStateChanges } from "./core/action"
 export { Lambda, isArrayLike } from "./utils/utils"
 export { isComputingDerivation as _isComputingDerivation } from "./core/derivation"
-// TOOD: move next thee to confgiure
-export { setReactionScheduler as _setReactionScheduler, onReactionError } from "./core/reaction"
-export { reserveArrayBuffer as _reserveArrayBuffer /* See #734 */ } from "./types/observablearray"
+export { onReactionError } from "./core/reaction"
 export { interceptReads as _interceptReads } from "./api/intercept-read"
 export { IComputedValueOptions } from "./core/computedvalue"